move router.ts
